[ASAP-1683] - extract MetricsCard from TeamLeadershipMetrics component
